Return dicts with agent-specific information (agent_info) from policy(). This will be useful for implementing our new version of AWR, and possibly other algorithms.
To enable that, introduced two new nested ops: nested_zip and nested_stack, allowing to stack together arbitrary pytrees, and used them to refactor the trajectory batching/padding code so it works on arbitrary pytrees passed as agent_info. This code can be simplified further, I'm leaving that for the future.

PiperOrigin-RevId: 310198476